requirements.txt
sqlalchemy>=2.0.0
asyncpg>=0.29.0
bcrypt>=4.0.0
defusedxml>=0.7.1
python-dotenv>=1.0.0
alembic>=1.13.0
psutil>=5.9.0

